+++ Tests/MapboxMapsTests/Foundation/Camera/BasicCameraAnimatorTests.swift
@@ -57,6 +57,16 @@ final class BasicCameraAnimatorTests: XCTestCase {
         XCTAssertTrue(propertyAnimator.finishAnimationStub.invocations.isEmpty)
     }
 
+    func testIsReversed() {
+        animator.isReversed = true
+
+        XCTAssertEqual(propertyAnimator.setIsReversedStub.parameters, [true])
+
+        animator.isReversed = false
+
+        XCTAssertEqual(propertyAnimator.setIsReversedStub.parameters, [true, false])
+    }
+
     func testStartAndStopAnimation() {
         animator.addAnimations { (transition) in
             transition.zoom.toValue = cameraOptionsTestValue.zoom!
@@ -66,6 +76,7 @@ final class BasicCameraAnimatorTests: XCTestCase {
 
         XCTAssertEqual(propertyAnimator.startAnimationStub.invocations.count, 1)
         XCTAssertEqual(propertyAnimator.addAnimationsStub.invocations.count, 1)
+        XCTAssertEqual(propertyAnimator.addCompletionStub.invocations.count, 1)
         XCTAssertNotNil(animator?.transition)
         XCTAssertEqual(animator?.transition?.toCameraOptions.zoom, 10)
 
@@ -73,5 +84,41 @@ final class BasicCameraAnimatorTests: XCTestCase {
         XCTAssertEqual(propertyAnimator.stopAnimationStub.invocations.count, 1)
         XCTAssertEqual(propertyAnimator.finishAnimationStub.invocations.count, 1)
         XCTAssertEqual(propertyAnimator.finishAnimationStub.invocations.first?.parameters, .current)
+    }
+
+    func testAnimatorCompletionUpdatesCameraIfAnimationCompletedAtEnd() throws {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraOptionsTestValue.zoom!
+        }
+        animator.startAnimation()
+        let completion = try XCTUnwrap(propertyAnimator.addCompletionStub.parameters.first)
+
+        completion(.end)
+
+        XCTAssertEqual(mapboxMap.setCameraStub.invocations.count, 1)
+    }
+
+    func testAnimatorCompletionUpdatesCameraIfAnimationCompletedAtStart() throws {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraOptionsTestValue.zoom!
+        }
+        animator.startAnimation()
+        let completion = try XCTUnwrap(propertyAnimator.addCompletionStub.parameters.first)
+
+        completion(.start)
+
+        XCTAssertEqual(mapboxMap.setCameraStub.invocations.count, 1)
+    }
+
+    func testAnimatorCompletionDoesNotUpdateCameraIfAnimationCompletedAtCurrent() throws {
+        animator.addAnimations { (transition) in
+            transition.zoom.toValue = cameraOptionsTestValue.zoom!
+        }
+        animator.startAnimation()
+        let completion = try XCTUnwrap(propertyAnimator.addCompletionStub.parameters.first)
+
+        completion(.current)
+
+        XCTAssertEqual(mapboxMap.setCameraStub.invocations.count, 0)
     }
 }