+++ Tests/MapboxMapsTests/Foundation/MapboxMapTests.swift
@@ -144,8 +144,6 @@ final class MapboxMapTests: XCTestCase {
         let map = MapboxMap(mapClient: MockMapClient(), mapInitOptions: MapInitOptions())
 
         // Compilation check only
-        _ = map as MapTransformDelegate
-        _ = map as CameraManagerProtocol
         _ = map as MapFeatureQueryable
         _ = map as ObservableProtocol
         _ = map as MapEventsObservable